pyproject.toml
[tool.poetry]
name = "sudoku-solver"
version = "0.1.0"
description = "Solve Sudoku with AI because solving as a human is hard"
authors = ["David Der <david.der@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.32.0"
opencv-python = "^4.9.0"
numpy = "^1.26.0"
openai = "^1.12.0"
pillow = "^10.2.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"